+++ src/Loader/YamlFileLoader.php
@@ -19,6 +19,7 @@ class YamlFileLoader extends CompatibilityFileLoader
         'resource',
         'type',
         'channel',
+        'pattern',
         'handler',
         'callback',
         'defaults',
@@ -137,13 +138,19 @@ class YamlFileLoader extends CompatibilityFileLoader
             }
         }
 
+        if (isset($config['pattern'])) {
+            $pattern = $config['pattern'];
+        } else {
+            $pattern = $config['channel'];
+        }
+
         if (isset($config['callback'])) {
             $callback = $config['callback'];
         } else {
             $callback = $config['handler'];
         }
 
-        $route = new Route($config['channel'], $callback);
+        $route = new Route($pattern, $callback);
         $route->addDefaults($defaults);
         $route->addRequirements($requirements);
         $route->addOptions($options);
@@ -165,16 +172,20 @@ class YamlFileLoader extends CompatibilityFileLoader
         }
 
         if (isset($config['resource'])) {
-            if (isset($config['channel'])) {
-                throw new \InvalidArgumentException(sprintf('The routing file "%s" must not specify both the "resource" key and the "channel" key for "%s". Choose between an import and a route definition.', $path, $name));
+            if (isset($config['channel']) || isset($config['pattern'])) {
+                throw new \InvalidArgumentException(sprintf('The routing file "%s" must not specify both the "resource" key and the "pattern" key for "%s". Choose between an import and a route definition.', $path, $name));
             }
         } else {
             if (isset($config['type'])) {
                 throw new \InvalidArgumentException(sprintf('The "type" key for the route definition "%s" in "%s" is unsupported. It is only available for imports in combination with the "resource" key.', $name, $path));
             }
 
-            if (!isset($config['channel'])) {
-                throw new \InvalidArgumentException(sprintf('You must define a "channel" for the route "%s" in file "%s".', $name, $path));
+            if (isset($config['pattern']) && isset($config['channel'])) {
+                throw new \InvalidArgumentException(sprintf('The route "%s" in file "%s" must not specify both the "channel" key and the "pattern" key.', $name, $path));
+            } elseif (isset($config['channel'])) {
+                trigger_deprecation('gos/pubsub-router-bundle', '2.4', 'Setting the "channel" key for the route "%s" in file "%s" is deprecated and will not be supported in 3.0, use the "pattern" key instead.', $name, $path);
+            } elseif (!isset($config['pattern'])) {
+                throw new \InvalidArgumentException(sprintf('You must define a "pattern" for the route "%s" in file "%s".', $name, $path));
             }
 
             if (isset($config['callback']) && isset($config['handler'])) {